DesktopAlertDialog scrim color
https://github.com/JetBrains/compose-jb/issues/2580
Also, add temporary PopupDialog. In the future it should be changed to LayerDialog.(https://github.com/JetBrains/compose-jb/issues/933)
